+++ tests/subsys/zigbee/osif/serial/serial_basic_api/CMakeLists.txt
@@ -15,6 +15,7 @@ add_definitions(-Wno-packed-bitfield-compat)
 zephyr_compile_definitions(CONFIG_ZB_HAVE_SERIAL)
 zephyr_compile_definitions(CONFIG_ZIGBEE_HAVE_SERIAL)
 zephyr_compile_definitions(CONFIG_ZIGBEE_UART_DEVICE_NAME="UART_1")
+zephyr_compile_definitions(CONFIG_ZBOSS_TRACE_BINARY_NCP_TRANSPORT_LOGGING)
 zephyr_compile_definitions(CONFIG_ZIGBEE_UART_TX_TIMEOUT=10)
 zephyr_compile_definitions(CONFIG_ZIGBEE_UART_RX_TIMEOUT=10)
 zephyr_compile_definitions(CONFIG_ZIGBEE_UART_PARTIAL_RX_TIMEOUT=10)
