+++ tests/subsys/zigbee/osif/serial/serial_via_logger/prj.conf
@@ -5,3 +5,4 @@
 #
 
 CONFIG_ZTEST=y
+CONFIG_RING_BUFFER=y
